Fix: Partial Agg may cause unpin twice
ERROR:  cannot unpin a segment that is not pinned (dsm.c:983)  (seg2 slice17 127.0.1.1:7004 pid=7331) (dsm.c:983)

In scan path, if have not `delete` operator in target list, Then should not build block bitmap.

GetBlockNumber is not a process-safe function. Should not call it when caller not in update or delete operator.