Store transformed PartBoundSpec in pg_pax_tables
Constant values are not directly saved in PartitionBoundSpec.
For range partitions, the lowerdatums and upperdatums save
the list of PartitionRangeDatums, not Const. The transform
happens after parser, but we transform directly here.

Now, the partitionspec in pg_pax_tables are completely the same
as normal pg_class.relpartbound for partition tables.