PAX: Add new guc pax.enable_filter
PAX always calculates min/max values in the micro partitions,
so we could apply these values to filter the whole micro partition.
There is no way to bypass the filtering process.

However, there's a case that we need to disable the filtering,
like BUGs.

The guc is on by default. We should manually turn it off.