Pax: add MicroPartitionWriter merging logic back to optimize performance
When pax uses TableParitionWriter to insert data, the file merging operation will be performed
in the Close.

If the current table already build the indexes, Pax can only use delete-update to update the
indexes.

If the current table is not indexed, pax assumes that the CTID in the TableTupleSlot(using
SetBlockNumber + SetTupleOffset) has no meaning. which means that we can directly call
writer->MergeTo to speed up the merge process.