Fix: the index in pax aux table should be same as aux table
The pax aux table will used the oid(7064) as namespace which is one of system table namespaces.

After create a pax table, the pax aux table will set the table space to invalid oid(0),
no matter currently table space is pg_default or not. This is because the table space of the
pax aux table is unchangeable (system table permission).

Whenever an pax aux table is created, the index of the pax aux table is also created. But
the index of the aux table is set to the same table space as the pax table
(current table space). Then the user can not successfully move the index to another table space.