Pax change the default file split strategy to guc
In TPCH or TPCDS, If pax uses the default file splitting configuration,
it will make the data file too small and affect the reading performance.

But if it is in other case, using current file splitting configuration may
not be a problem.

Current change make the file splitting configuration to a GUC value, which
allows users to configure different configurations according to different case,
and the default value of file splitting configuration changed to a larger value.

Also, Pax will no longer set a default encoding for columns.