Fix: pax custom object support some of event_trigger callback
The CustomObjectClass in pax have not implements the callback object_type_desc
 and object_identity_parts which used in trigger call.

If current event_trigger registered, then CBDB can't drop the pax table.
Because the event_trigger try get the object description when table dropped.
But pax does not implement the callback function, so an error will be raise during
this process.