Use scan_flags to indicate whether the table AM support vectorization
Currently, we use AM handler's oid to judge whether the AM support
vectorization, generate the vectorization plan if AM does so.

For those new AMs defined in plugins, we may need to access
the oid through the plugin's interface, it is not convenient
because of the dependency. We use AM's callback 'scan_flags'
instead of AM, and each new AM in plugins only need to return
the flags represented the supported features.