Pax: TupleDesc in TableTupleSlot is no longer used
In am function tuple_insert, PAX will use the TupleDesc which in TableTupleSlot
to build the PORC format column.

But the TupleDesc in TableTupleSlot may different with TupleDesc in Relation. In fact,
TupleDesc in TableTupleSlot is incomplete, it is returned from the Result node, and
the attstorage is queried from pg_type.

Current change replace the TupleDesc in TableTupleSlot with TupleDesc in Relation, then
PAX can get a accurate TupleDesc.

The reason why scan_getnextslot is not changed is that in the *initnode stage, TupleDesc
 in TableTupleSlot is assigned relation->rd_att, and the two are consistent.