Op: PAX Precompute null counts array in GetTuple
In GetTuple, if multiple tuple are fetched, PAX will call CountNulls multiple times.
The method CountNulls will do foreach and accumulate the number of NULL.

In the current change, PAX will pre-calculate the number of null values for each position,
so that only one for-each is needed when we call the GetTuple.