PAX: enable kind of UDF to dump the PAX file
In some customer environments, we may not be allowed to get/access the customer's
data files, or even use the shell.

Therefore, the current changes also support the use of UDF to dump user data.
At the same time, UDF can connect to object storage to support storage_am debugging
which in not support in pax_dump.