PAX: Support bloom filter in the statistics
In the customer case: customer will use the varchar store as enum(few set of value in the column).

So in current changes, bloom filter is supported in PAX. The enum case will supporbe ted as a subset of bloom filter. bloom filter can also do more filter.

In addition, the current changes support filtering of ScalarArrayOpExpr, currently only bloom filter is supported for processing ScalarArrayOpExpr. There are plans to handle different *opexpr with supported filter types in PAX.