PAX: Add wal log
Based on Custom WAL Resourcr Manager, Pax registers a custom RmgrData structure
with RMGR ID=199 to implement a custom WAL Log.

Currently, 3 types of wal log are defined which used for create dir, truncate
dir and write binary file.
In the mirror redo process, we add two xlog hook functions to check whether
there is invalid xlog for file and directory writing.
pg_waldump can be used to dump the xlog of pax if it's build with option
--enable_pax.

Future Optimization:
    After implementing wal log, we can avoid writing small insert to disk and
build a new memory structure (such as an ordered LSM Tree). When the threshold
is reached, the data is written to disk, the data in the pax file is ordered,
so that index information can be added to the file later.