PAXPY: Changed the CmakeLists
PAXPY relies on CBDB install before it can be built.

In the current changes, PAXPY no longer relies on CBDB install before
building, but directly links storageformat.so.However, since PAXPY is
not tested in CI, PAXPY may not have changed when some APIs of PAX
itself are changed. Consider adding it to CI later.