PAX: remove the GUC VECTOR_MAX_BATCH_SIZE_GUC_NAME
In the current change, vec.max_batch is no longer use to determine the number of rows returned
by record batch.

But for the class VecAdapter, the range interface is still retained. For the PAX, the cost of
splitting by range is small. If the range interface of the class VecAdapter is no longer needed
in the future, the related interface parameters will also be removed.